yoouk/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
Django==5.2.3
django-allauth==65.9.0
gTTS==2.5.4
idna==3.10
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
PyPDF2==3.0.1
pytesseract==0.3.13
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
